Fixed issue with output order of 'list-packages'
The output was previously sorted (alphabetically) by the name of the MPR 
package that any given package was built from (either 'pkgbase' or the 
first item in 'pkgname'). The output is now sorted by the name of the 
built package instead (the actual value of 'pkgname' for the built 
package).